Move feedback from submissions to intro mustache
Moved existing_feedback_from_teachers to support mustache output
