+++ templates/intro.mustache
@@ -81,5 +81,10 @@
 
 <hr>
 
+{{! Feedback }}
+{{#feedback}}
+    {{{.}}}
+{{/feedback}}
+
 {{! Test data output. }}
 {{{table}}}
