+++ templates/submission.mustache
@@ -95,13 +95,6 @@
         </div>
     {{/final}}
 
-    {{! TODO - Feedback - probably should be somewhere else? }}
-    {{#feedback}}
-        <hr>
-        <h1>What is this?</h1>
-        {{{.}}}
-    {{/feedback}}
-
     {{/cansubmit}}
 </div>
